chore: move upload/download to actions drawer
so that we can fit back/forward buttons on the toolbar
