+++ src/web/topic/components/Surface/MoreActionsDrawer.tsx
@@ -1,4 +1,4 @@
-import { AutoStoriesOutlined, Build, Close, Route } from "@mui/icons-material";
+import { AutoStoriesOutlined, Build, Close, Download, Route, Upload } from "@mui/icons-material";
 import {
   Divider,
   Drawer,
@@ -9,21 +9,66 @@ import {
   ListItemText,
   ToggleButton,
 } from "@mui/material";
+import fileDownload from "js-file-download";
+import { StorageValue } from "zustand/middleware";
 
+import { errorWithData } from "../../../../common/errorHandling";
 import { Perspectives } from "../../../view/components/Perspectives/Perspectives";
-import { useIsTableActive, useShowImpliedEdges } from "../../store/store";
+import { migrate } from "../../store/migrate";
+import { TopicStoreState, useIsTableActive, useShowImpliedEdges } from "../../store/store";
 import { useOnPlayground } from "../../store/topicHooks";
-import { resetTopicData } from "../../store/utilActions";
+import { getPersistState, resetTopicData, setTopicData } from "../../store/utilActions";
+import { getTopicTitle } from "../../store/utils";
 import { toggleShowImpliedEdges } from "../../store/viewActions";
+
+// TODO: might be useful to have downloaded state be more human editable;
+// for this, probably should prettify the JSON, and remove position values (we can re-layout on import)
+const downloadTopic = () => {
+  const persistState = getPersistState();
+
+  const topicState = persistState.state;
+  const topicTitle = getTopicTitle(topicState);
+  const sanitizedFileName = topicTitle.replace(/[^a-z0-9]/gi, "_").toLowerCase(); // thanks https://stackoverflow.com/a/8485137
+
+  fileDownload(JSON.stringify(persistState), `${sanitizedFileName}.json`);
+};
+
+const uploadTopic = (event: React.ChangeEvent<HTMLInputElement>, sessionUsername?: string) => {
+  if (event.target.files === null) return;
+
+  const file = event.target.files[0];
+  if (!file) return;
+
+  file
+    .text()
+    .then((text) => {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- TODO: validate that JSON matches interface
+      const persistState = JSON.parse(text) as StorageValue<TopicStoreState>;
+      if (!persistState.version) {
+        throw errorWithData("No version found in file, cannot migrate old state", persistState);
+      }
+
+      const migratedState = migrate(persistState.state, persistState.version) as TopicStoreState;
+
+      setTopicData(migratedState, sessionUsername);
+    })
+    .catch((error) => {
+      throw error;
+    });
+};
 
 interface Props {
   isMoreActionsDrawerOpen: boolean;
   setIsMoreActionsDrawerOpen: (isOpen: boolean) => void;
+  sessionUser?: { username: string } | null;
+  userCanEditTopicData: boolean;
 }
 
 export const MoreActionsDrawer = ({
   isMoreActionsDrawerOpen,
   setIsMoreActionsDrawerOpen,
+  sessionUser,
+  userCanEditTopicData,
 }: Props) => {
   const onPlayground = useOnPlayground();
   const isTableActive = useIsTableActive();
@@ -57,9 +102,31 @@ export const MoreActionsDrawer = ({
         <Divider />
 
         <ListItem disablePadding={false}>
-          <IconButton color="inherit" title="Reset" aria-label="Reset" onClick={resetTopicData}>
-            <AutoStoriesOutlined />
+          <IconButton
+            color="inherit"
+            title="Download"
+            aria-label="Download"
+            onClick={downloadTopic}
+          >
+            <Download />
           </IconButton>
+
+          {userCanEditTopicData && (
+            <>
+              <IconButton color="inherit" component="label" title="Upload" aria-label="Upload">
+                <Upload />
+                <input
+                  hidden
+                  accept=".json"
+                  type="file"
+                  onChange={(event) => uploadTopic(event, sessionUser?.username)}
+                />
+              </IconButton>
+              <IconButton color="inherit" title="Reset" aria-label="Reset" onClick={resetTopicData}>
+                <AutoStoriesOutlined />
+              </IconButton>
+            </>
+          )}
 
           {!isTableActive && (
             <>
